+++ src/net/Qt.NetCore.Sandbox/Program.Tests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Qt.NetCore.Types;
 
 namespace Qt.NetCore.Sandbox
 {
