+++ src/net/Qt.NetCore.Tests/Types/NetTypeManagerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using FluentAssertions;
+using Qt.NetCore.Types;
 using Xunit;
 
 namespace Qt.NetCore.Tests.Types
