+++ src/net/Qt.NetCore/Callbacks.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using AdvancedDLSupport;
+using Qt.NetCore.Types;
 
 namespace Qt.NetCore
 {
