+++ src/net/Qt.NetCore/Types/NetTypeInfo.cs
@@ -3,7 +3,7 @@ using System.Runtime.InteropServices;
 using AdvancedDLSupport;
 using Qt.NetCore.Internal;
 
-namespace Qt.NetCore
+namespace Qt.NetCore.Types
 {
     public class NetTypeInfo : BaseDisposable
     {
